+++ extensions/warp/CHANGELOG.md
@@ -7,3 +7,7 @@
 - "Open Launch Configuration" command
 - "New Tab" command
 - "New Window" command
+
+## [Warp Workflows] - 2023-08-11
+
+- "Open Workflow" command